Remove obsolete Go-version-based alternatives
Since utils requires Go 1.23, code specifically for older versions is
no longer necessary.
