+++ exec/exec.go
@@ -18,6 +18,7 @@ package exec
 
 import (
 	"context"
+	"errors"
 	"io"
 	"io/fs"
 	osexec "os/exec"
@@ -95,6 +96,21 @@ type executor struct{}
 // New returns a new Interface which will os/exec to run commands.
 func New() Interface {
 	return &executor{}
+}
+
+// maskErrDotCmd reverts the behavior of osexec.Cmd to what it was before go1.19
+// specifically set the Err field to nil (LookPath returns a new error when the file
+// is resolved to the current directory.
+func maskErrDotCmd(cmd *osexec.Cmd) *osexec.Cmd {
+	cmd.Err = maskErrDot(cmd.Err)
+	return cmd
+}
+
+func maskErrDot(err error) error {
+	if err != nil && errors.Is(err, osexec.ErrDot) {
+		return nil
+	}
+	return err
 }
 
 // Command is part of the Interface interface.
